Constraints pt 2

See merge request jwzp_kp_nk/jwzp_kp_nk!8

+++ src/main/java/uj/jwzp/kpnk/GymApp/service/CoachService.java
@@ -2,20 +2,22 @@ package uj.jwzp.kpnk.GymApp.service;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import uj.jwzp.kpnk.GymApp.exception.coach.CoachNotFoundException;
+import uj.jwzp.kpnk.GymApp.exception.coach.AssignedEventsException;
 import uj.jwzp.kpnk.GymApp.model.Coach;
 import uj.jwzp.kpnk.GymApp.repository.CoachRepository;
 
-import java.util.Optional;
 import java.util.Set;
 
 @Service
 public class CoachService {
 
     private final CoachRepository repository;
+    private final EventService eventService;
 
     @Autowired
-    public CoachService(CoachRepository repository) {
+    public CoachService(CoachRepository repository, EventService eventService) {
         this.repository = repository;
+        this.eventService = eventService;
     }
 
     public Set<Coach> allCoaches() {
@@ -39,7 +41,7 @@ public class CoachService {
 
     public void removeCoach(int id) {
         if (repository.coach(id).isEmpty()) throw new CoachNotFoundException(id);
-
+        if (!eventService.eventsByCoach(id).isEmpty()) throw new AssignedEventsException(id);
         repository.removeCoach(id);
     }
 }
